EDU-1034 Validate placeholders
In old solutions on Stepik we didn't have placeholder tags, so we need to validate them.
Now if we don't see placeholder tags in solution text we set placeholder length and offset
to -1. In validate methods we check it, and if one of placeholders is invalid we don't check
task and show notification in editor