Init course on post
On post we store posted course in StudyTaskManager, but if we don't call
init lessons contain a link to the old instance of a course. It causes problems,
for example, when lesson is moved: we get source course from lesson and target
one from StudyTaskManager, so in fact, the lesson is deleted from the old instance
and added to the new one, so we get a course with wrong number of items in
StudyTaskManager.

(cherry picked from commit 8da63a3)