EDU-1963 Exception on "Open as Educator" Django Tutorial course
The problem is that we need the whole course structure to save all config
files. Than I figured out, that the call `saveAll` do something meaningful
in only one specific case: user created a project after yaml feature was enabled.
So it's not a big deal if they will be created a bit later after project
is opened, so I added a check and then, as EDU_PROJECT_CREATED flag is
set after project is opened I wrapped this call into `registerPostStartupActivity`