EDU-2183: fix check action for AtomicKotlin in Android Studio
For one module kotlin courses we modified some run configuration params
if configuration was gradle configuration. If configuration wasn't gradle one, we returned null
Turned out that AS produces own configurations (not gradle based) so we always returned null in AS.
Also, AS configurations work fine even with one-module course.

So this change just return unmodified configuration in AS