EDU-2280: fix navigation tests on CLion
We create each course file in separate write action that leads that some other code can be invoked between the creation.
Including `CourseViewUtils.hasVisibleTaskFilesNotInsideSourceDir`.

And in case of incomplete course structure on file system, it's possible not to find file virtual file for some task file.
Previously, we called `Logger#error` in such sutiations. But turned out that `TestLogger` (logger implementation in test)
throws an exception in `TestLogger#error` that's a cause of tests fail.

This change just use `warn` instead of `error`