EDU-2338: correctly set up gradle module names with parenthesis
Previously, we didn't replace parenthesis with `_` in `settings.gradle` but do it in `GeneratorUtils#sanitizeName`.
It led to invalid command name in output and theory task checkers.
Also, if task contains code that cannot be compiled and the corresponding module contains parenthesis in the name,
gradle produces weird error message (not related to compilation)