EDU-2763 support GoLand, fix bug with default language
While creating new course, we selected default language
based on current IDE.
Invoking setSelectedItem forced to work corresponding listener, and
we filled all other course info there (name, description, etc)

The problem was that Go is the first language in the list,
and it is selected by default. When we select it again, nothing happens
and our listener is not invoked. Now we call it by ourselves,
and only then we register the listener.