EDU-2763 cleanup
There could be the scenario when defaultCourseType == null, and
myCourseDataComboBox.selectedItem != null
(first item is selected by default), and I break this scenario

But it's a kind of obscure situation, we always should have not null
defaultCourseType. If shit happened, user anyway will want
to select language, and after selection everything will be
filled by the ItemListener.