EDU-2843 Fix painting placeholders if file contains tabs
We had this issue because we constructed LogicalPosition by ourselves.
It should take into account size of the symbol (e.g. tab), and we
did not fix it.
Now we are taking LogicalPosition from offset and we don't need to
think about tabs - it was made for us in `offsetToLogicalPosition()`