EDU-2921 Refactor hyperskill configurators
The main problem is that we have too many duplicated code in Hyperskill configurators,
and we do not have a possibility to reuse it.
We decided to make more structured ways of inheritance & delegation,
it created many problems for us with converting some files to Kotlin
(that's the only way to make implementation by delegation work),
but finally it looks that everything works.