EDU-2673: initial support for in course links
In general, it's a local file-system path relative to project directory in educator mode with `course://` schema. Path should point to some course item: section, lesson, task or task file.
Also, if path contains symbols invalid in uri, they should be encoded

There is a special support for framework lessons where paths in student mode differ. But currently, it's not possible to navigate to non current task in framework lesson