EDU-4052 explicitly set redirect file for gradle configurations
For java there is a list of run configurations that are supposed to have redirect input from file. If a configuration id is not in the list, redirect input is ignored.
To solve this problem codeforces configurations used to have the same id as Application run configurations which led to Application run configurations corruption and possibly other problems.