EDU-4444: drop `EduBundle` in android module
It was introduced for old versions of AS that don't have `DynamicBundle`. Now, the oldest support platform version is 203 and AS based on 203 platform has `DynamicBundle` class so this class is redundant now.
Also, it may break plugin when we start merging all plugin jars into single one to support new plugin model