EDU-5404: Pass the task into absolutizePaths()
The problem was that at the initial loading of the project, we may want to show a description of first task, but we haven't opened any tasks yet. Because of that getCurrentTask() is null and the paths are incorrectly absoluted. So I pass the desired task directly into the absolutizePaths function.