EDU-2796, EDU-5191: handle copy event, handle creation of a new config
Now if a new -info.yaml file is created, either by itself or by a copy operation, we try to add it to the parent config. After that, we refresh the project with the reason that the course structure changed.

Fixed the comparator that tries to process configs in a certain order: first course configs, then sections configs, then lessons configs, then task configs.

Fixed testing of files to be inside folders: `a.txt` is no more treated as a child of the `a` folder.