Disable code instrumentation by gradle-intellij-plugin
Since 1.13, gradle-intellij-plugin changed a way how instrumented code is added to classpath. But developing significantly got worse in multi module plugins after such changes because from IDE point of view, module started depend not on other modules themselves but on the corresponding instrumented jar artifacts. As a result, it affects name resolution which now navigates not to source code but to compiled instrumented classes which works bad in case of Kotlin.

But actually, we don't need instrumentation now since we got rid of all .form files, the project contains less than 100 java files, and we almost don't produce new Java code, i.e. we almost doesn't have code which are supposed to be processed by instrumentation.
At the same time, we want to update gradle-intellij-plugin because it contains fixes/features related to platform changes

So this change disables code instrumentation