Save `problemIndex` in codeforces tasks
It's a string-based index of the task which is used in Codeforces urls.
It is mostly used for submissions and different links.

Previously, the plugin tried to match submissions task names, but it found out that a task name in a contest project may differ from name in a submission object:
- non-english language of context
- different task name wording in a problem and contest pages
- technical task name for some languages (always for C++, Rust, in case of special symbols in Gradle-based projects)
After these changes, we always use `problemIndex` to match submissions with tasks and calculated task-related urls

Fixes EDU-4847, EDU-5014, EDU-5439, EDU-5869, EDU-5870
Also, probably fixes EDU-4852, EDU-5434