EDU-6284 description is empty for the first task in newly installed IDE
Valentina Kiryushkina
Commented just now

For dev, why it happens.
The problem is that we are getting 0 as a font size in com.jetbrains.edu.learning.taskToolWindow.ui.styleManagers.TypographyManager#uiSettingsFontSize
It happens because UISettings.getInstance().fontSize is zero for the newly created course. fontSize is a not-roamable property that is stored in other.xml (see com.intellij.ide.ui.NotRoamableUiSettings). After some refactorings in 232, it stopped returning default font