EDU-6290 Support segmented button in `SubmissionsTab`
As a part of the Solution Sharing feature, the `SubmissionsTab` has been extended with an additional panel and a segmented button to manage it. All changes are under the `Registry`. Note that some UI features from the mockups (such as the scroll separator, icons, etc.) will be added in the next commits.