EDU-6281: initialize `solutionsSharing` property in `MarketplaceSettings`
Since `MarketplaceSettings` is an application-level service, its instance can be called from anywhere in the application and exists regardless of whether the course is open. Hence, obtaining and updating the user's preference for Solutions Sharing is now implemented in this class. This allows the current state of this property to be retrieved from other parts of the application where it is needed (e.g. `MarketplaceOptions` and `ShareMySolutionsAction`) without producing additional requests.