EDU-6452 Enable Solution Sharing Inline Banners
Implement Solution Sharing Inline Banners to prompt users to enable this feature upon correctly solving a task for the first time. If rejected, users will be prompted again after some delay. The duration of the delay is managed by the registry value for easier testing with a default value of 14 days.

Note: Inline Banners have been available in the platform only since 233; for older versions (231 & 232), notifications with actions are displayed instead.