EDU-6503: properly validate tool window links in headless mode
Links validation is supposed to work correctly in headless mode (on CI, for example). But IDEs don't load tool windows in headless mode by default. As a result, you can't just use `ToolWindowManager#getToolWindow` to ensure if a tool window link is valid or not

These changes add manual checks of the corresponding tool window extensions