EDU-6607: fix binary incompatibility with the latest versions of Rust plugin
The latest versions of `Rust` plugin changed the signature of `CargoProjectsService.refreshAllProjects` method - new argument was added.
As a result, it led to binary incompatibility, `java.lang.NoSuchMethodError: 'java.util.concurrent.CompletableFuture org.rust.cargo.project.model.CargoProjectsService.refreshAllProjects()` in particular.

These changes update versions of rust plugin which are used for compilation to be compatible with the latest changes in Rust plugin