EDU-6489: Don't Throw Error on Failed Network Call
The investigation of the reported exception revealed that it arises due to client issues (e.g., an interrupted internet connection). There is no evidence that an error is received from the remote server. Therefore, we decided to log the error instead of throwing an exception, and return an empty list of solution files.