EDU-6442: Clear CCFrameworkLessonRecordStorage after each edu test
The problem is:
When calling tests (all tests for course creator, not just class `CCFrameworkLessonRecordStorageRenameTest`), the first test called from `CCFrameworkLessonRecordStorageRenameTest` fails.

This is due to other tests that used record storage, because it was not cleared after them.
In this commit, I added a storage cleanup after each test.