EDU-6760: allow C++ course creation if new C++ language engine (AKA Clion Nova) is enabled
Previously, the corresponding plugin module depended on `com.intellij.cidr.lang` because we used classic C++ PSI.
But if the new C++ language engine is enabled, `com.intellij.cidr.lang` is disabled. As a result, our integration was disabled as well.

These changes drop explicit usage of classic C++ PSI not to depend on any engine which allows us to work regardless of what engine is enabled