EDU-6752: YAML mapper for EduFiles
EduFiles now also have their mixin and builder to be written and read from YAML. They are base classes for corresponding TaskFiles mixin and builder.
EduFiles in YAML have only the name field because other features such as isVisible are not implemented for additional files.
EduFile has a special mixin and builder for the case of the student mode. In this mode, the is_binary field is added. The same way as for the TaskFiles: we don't store the binarity for a teacher.