EDU-6752: Collect additional files for courses if there are no in YAML
When the project is being opened and has YAML with an empty additional_files section, it is considered to be an old version of the project, and additional files are automatically collected to the YAML.
Additionally, new courses are created with empty YAML, so that their additional files are collected.
It is better if the CourseProjectGenerator created a list of additional files by itself, so one did not need to collect them during the course creation, but this is not easy to rewrite.