EDU-7029 Slow operation: EduStartupActivity.setupProject
Added functionality to refresh course project directory files in the background.
Moving `VfsUtil.markDirtyAndRefresh` to a background thread is necessary because this operation can be time-consuming, as evidenced by exceptions indicating prolonged execution times.
Since `VfsUtil.markDirtyAndRefresh` can be executed on the EDT but is also suitable for a background thread (BGT) without a read action, performing it in the background should help prevent potential UI freezes and improve overall performance.