EDU-6980: make `afterProjectOpened` work for projects from CourseStorage
Fix for the case, when the course is not found in recent projects, but is present in the CourseStorage. This happens when a user works with some edu project, then closes it, removes from the list of recent projects (it still stays in CourseStorage, one can't just remove a course from the CourseStorage). And launches the same course from the browser.