EDU-7505: use JCEF in Task description if possible
Previously, it was possible to have disabled and enabled JCEF during the first and the second IDE runs. As a result, the plugin used Swing because this value was saved in settings in the first run.

Now, `EduSettings` tries to switch value to the best possible option (JCEF is better than Swing) in the current IDE run preserving manual user choice if possible