EDU-7526: do not fail if rem dev service started twice
Previously, we used an assumption that in rem dev environment there is only one project, and it cannot be reopened. And according to this assumption, we started some application wide service and allowed it to start only once.

It seems it's not true that project in remote environment cannot be reopened. So, after these changes, the plugin won't fail because of attempt to launch application wide processing second time