EDU-7585: use declarative way to fill right action toolbar in CheckPanel
Previously, necessary actions were hardcoded in CheckPanel itself. It not only made the component harder to customize, it also added unnecessary dependency in Marketplace code

Now, we use `Educational.CheckPanel.Right` group defined in plugin manifest to create the corresponding list of actions for the corresponding toolbar. At also allowed not to update set of actions manually as it was before