EDU-6988: avoid `Device is already activating` error
Previously, as a part of running run configuration by check action, we checked that `RunProfileState` is not null.

But in some cases (like Android Instrumentation tests) it invokes some code which is supposed to be called only once per run configuration launch. In combination with that `env.state == null` check wasn't actually useful, we've just dropped it