EDU-7716 Replace CourseTranslationDialog with CourseTranslationPopup
The commit removes CourseTranslationDialog and introduces a new CourseTranslationPopup to improve the user experience.
The changes also include rearranging import statements and modifying the action method in AITranslation.kt to accommodate for the introduced popup.
The new popup is more convenient and offers a better interaction for choosing the course translation language.
Deletion of the CourseTranslationDialog.kt file and creation of the CourseTranslationPopup.kt file form key parts of the changes.