Do not mark `TranslationSettings` as a light service
The service is not light since it's registered in the plugin manifest.
Moreover, it cannot be a light service since it's an application-level persistent service which uses non `RoamingType.DISABLED` roaming type. This type of services cannot be marked as a light service according to documentation https://plugins.jetbrains.com/docs/intellij/plugin-services.html#light-service-restrictions