EDU-6752: Update list of additional files on move
The set of additional files in the folder is preserved on move even if the folder contains files normally excluded by a configurator. For example, a directory might contain an additional file ".file" starting with the dot, and after moving the directory, this file will still be in the list of the additional files. Although such files are ignored by a configurator and do not go into the list of additional files automatically.