EDU-7872: Refactor Run Button logic
To detect run configurations, do not examine files on disk, but examine the RunnerAndConfigurationSettings#getPathIfStoredInArbitraryFileInProject value. It is stored in memory and is thus faster. Moreover, when a run configuration is changed, it is first changed in memory, and only after some time (sometimes, after a long time) it is serialized to disk. So the in-memory state is more relevant.