Implement connector for theory lookup & refactor errors
Since some of the translation errors are common with theory lookup, I've separated it into a common part.
The problem occurred when interacting with statistics, as the query requires an enum. In order not to spoil the statistics, I kept the old translation error format for interacting with statistics.

(cherry picked from commit 06e61bc1834ba7a7569371a4664998b6ed5cfd0f)