EDU-7746: Handle RemoteYamlLoadingException on archive creation
Reuse Course.loadRemoteInfoRecursively to load remote info for all items, including the course itself.
Because of reusing, some specific Hyperskill code was moved outside.

We also need to create YAML configs in the test, because reloading of remote info fails if an item has some ID, but its remote config file is absent.