EDU-7929: Additional files on FS action and on YAML update
The `isExcluded()` is changed, it is currently used in two places: in deciding which files are automatically added to the list of additional files when they are created on disk. And in collecting additional files in YAML migration.

The second usage is not changed. The first usage now uses the archive inclusion policy.