EDU-7940: Warn About Declined User Agreement in Editor Notification
When JetBrains Academy Plugin User Agreement is declined, we don't show any UI related to the Educational course. This may confuse users as they expect the course to appear. With this commit, we show a warning editor notification to let them know about the educational course absence due to declined User Agreement.