Fix rename study item tests
Starting with IJPGP 2.3, it loads all bundled plugins during the tests. As a result, it leads to new extensions in tests and some unexpected new behavior.
In this particular case, it turned out that our `moveHandler` can be not the first items in the list in move handlers because of handlers from Kotlin plugin, which breaks move behavior. These changes fix this order