Logging for Cognifire Study
Lifted Base class as abstract, deleted redundant data and the State manager for keeping this data, changed the isCodeChanged to use the CodeGenerator implementation.

Merged the logic for the Cognifire Logger and the Study Logger, also created the Manager for keeping the old data for future logs.

First commit, extracting the data for the code generation.


Merge-request: EDU-MR-4186
Merged-by: José Carlos Serize Portela <jose.serize.portela@jetbrains.com>