EDU-8244: fix `Check` action for output tasks in C++ courses
Previously, we assumed that CMake run configuration producer is the first producer in the list of `CidrTargetRunConfigurationProducer.getInstances(project)`.
But since 252 it's changed, and the plugin failed to find proper producer to create run configuration.

Now `CppCodeExecutor` doesn't rely on order and looks for proper run configuration producer based on executable targets