Reimplement C# tests with com.jetbrains.rider.test.BaseIntegrationTest
Tests for Rider should implement BaseIntegrationTest. Otherwise, they are not able to run because of a number of issues with the setup of the Rider environment.

Previously, some workarounds were used in CSharpTestBase.kt to bypass the incorrect environment, but now BaseIntegrationTest is the recommended solution.

It is impossible for the test class to be inherited from both the BaseIntegrationTest and the CourseGenerationTestBase, so some logic was extracted from CourseGenerationTestBase to the CourseGenerationTestBaseUtils interface.