EDU-8173: Allow interrupting Zhaba
The ZhabaComponent.stop() makes the animation stop after the current cycle is finished. The start() method returns `false` in this case.

We are going to stop the animation in case the window is resized or UI components are moved.

Also shorten 10-second animations so that zhaba will notice resize faster.