EDU-8534: Fix image loading, reload image on IDE zoom
The previous code used a common Swing method to load an SVG image, and it made its own computations of the image size, for example, it loaded a twice bigger image for a retina display. It was impossible to find out which magnification was used. This lad to a tricky interaction of actions to load an image / to draw an image / to convert image pixels to screen pixels. It seemed all of these actions were incorrect, but the errors somehow compensated for each other.

The new code uses the IntelliJ platform method of loading an SVG image that already loads it with a size corresponding to the current IDE zoom. This makes it possible to correctly convert between image pixels and screen pixels. But it is now important to reload images each time the IDE is zoomed.