Update jsoup to `1.21.2`
The platform now uses jsoup version 1.21.2. In out project, jsoup of the old version (`1.17.2`) is used and therefore many tests related to the editor's actions failed due to incompatibility.

Since jsoup formats the code a little differently in the new version, formatting of expected html in the tests was corrected.