EDU-8472: ZhabaGraph now stores transition animations
Before the fix we decided how to animate between transitions based on the pair of steps. But it is better when the Graph itself knows how to animate from one step to the next one.

If some steps are skipped (perform() returns null data for them): step1 -> skippedStep2 -> skippedStep3 -> step4, the transitionAnimator to be used is the one stored for the edge "skippedStep3 -> step4". But the StepData is passed to transitionAnimator is "stepData1" and "stepData4".