EDU-8472: Unify animations to jump down at the end
The Toad could jump away either with a sad image or with a happy one. Previously, we had separate ZhabaStep objects for both situations, separate ZhabaData objects for both situations, and two separate checks in Transition Animator. We are going to add one more jumping away, with the image where the Toad has a hat.

 The fix removes the duplication. Now there is only one ZhabaData class for jumping away. This data contains the image, either sad or happy.