EDU-8539: place JetBrains Academy actions in proper groups in Rider programmatically
Rider has its own action groups instead of default ones defined in other IDEs.
Previously, we solved the issue with unregistering and registering necessary actions again specifically for Rider.
But this way leads to a lot of code duplication. Moreover, each change in the original action group should be reflected in redefinition for Rider as well, which makes the solution quote fragile.

The new solution requires only a single definition of action groups from our side (which makes it much less error-prone)